requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
commonmark==0.9.1
ffmpeg-python==0.2.0
future==0.18.2
glob2==0.7
idna==3.3
Pillow==9.1.0
Pygments==2.12.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rich==12.3.0
six==1.16.0
tornado==6.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
